test: 유틸함수로 변경된 메소드 테스트코드경로 변경
