+++ src/controller/EventController.js
@@ -150,7 +150,6 @@ class EventController {
   }
 
   calculateDayEventDiscountAmount(menus, category) {
-    // utill
     const discountAmount = menus.reduce((amount, menu) => {
       if (category.includes(menu.name)) {
         return (amount += EVENT.DISCOUNT * menu.quantity);
@@ -159,26 +158,6 @@ class EventController {
     }, 0);
 
     return discountAmount;
-  }
-
-  convertAmountToString(amount) {
-    //utill
-    const stringAmount = String(amount);
-    const insertPoint = 3;
-    let countPoint = 0;
-    const convertToArray = [];
-
-    [...stringAmount].reverse().forEach((syllable) => {
-      convertToArray.push(syllable);
-      countPoint += 1;
-
-      if (countPoint === insertPoint) {
-        convertToArray.push(',');
-        countPoint = 0;
-      }
-    });
-
-    return convertToArray.reverse().join('');
   }
 }
